Refactor person_creator class
